refactor(communication/plugin): add snapshot test and correct naming
