+++ test/integration/editors/substation/bay-editor-wizarding.test.ts
@@ -1,13 +1,13 @@
 import { fixture, html, expect } from '@open-wc/testing';
 import fc from 'fast-check';
 
-import '../../mock-wizard.js';
-import { WizardingElement } from '../../../src/Wizarding.js';
+import '../../../mock-wizard.js';
+import { WizardingElement } from '../../../../src/Wizarding.js';
 
-import { getDocument } from '../../data.js';
-import { regexString, regExp } from '../../foundation.js';
+import { getDocument } from '../../../data.js';
+import { regexString, regExp } from '../../../foundation.js';
 
-describe('conducting-equipment-editor wizarding integration', () => {
+describe('bay-editor wizarding integration', () => {
   const doc = getDocument();
   let parent: WizardingElement;
 
@@ -15,16 +15,14 @@ describe('conducting-equipment-editor wizarding integration', () => {
     parent = <WizardingElement>(
       await fixture(
         html`<mock-wizard
-          ><conducting-equipment-editor
-            .element=${doc.querySelector('ConductingEquipment')}
-          ></conducting-equipment-editor
+          ><bay-editor .element=${doc.querySelector('Bay')}></bay-editor
         ></mock-wizard>`
       )
     );
 
     (<HTMLElement>(
       parent
-        ?.querySelector('conducting-equipment-editor')
+        ?.querySelector('bay-editor')
         ?.shadowRoot?.querySelector('mwc-icon-button[icon="edit"]')
     )).click();
     await parent.updateComplete;
@@ -53,34 +51,31 @@ describe('conducting-equipment-editor wizarding integration', () => {
       ).to.exist;
     });
   });
-  it('include 3 wizard inputs', async () => {
-    expect(parent.wizardUI.inputs.length).to.equal(3);
+  it('include 2 wizard inputs', async () => {
+    expect(parent.wizardUI.inputs.length).to.equal(2);
   });
-  it('the first input element only displaying the type', () => {
-    expect(parent.wizardUI.inputs[0]).to.have.property('disabled', true);
-  });
-  describe('the second input element', () => {
+  describe('the first input element', () => {
     it('edits the attribute name', async () => {
-      expect(parent.wizardUI.inputs[1].label).to.equal('name');
+      expect(parent.wizardUI.inputs[0].label).to.equal('name');
     });
     it('edits only for valid inputs', async () => {
       await fc.assert(
         fc.asyncProperty(regexString(regExp.tName, 1), async name => {
-          parent.wizardUI.inputs[1].value = name;
+          parent.wizardUI.inputs[0].value = name;
           await parent.updateComplete;
           expect(parent.wizardUI.inputs[0].checkValidity()).to.be.true;
         })
       );
     });
   });
-  describe('the third input element', () => {
+  describe('the second input element', () => {
     it('edits the attribute desc', async () => {
-      expect(parent.wizardUI.inputs[2].label).to.equal('desc');
+      expect(parent.wizardUI.inputs[1].label).to.equal('desc');
     });
     it('edits only for valid inputs', async () => {
       await fc.assert(
         fc.asyncProperty(regexString(regExp.desc), async desc => {
-          parent.wizardUI.inputs[2].value = desc;
+          parent.wizardUI.inputs[1].value = desc;
           await parent.updateComplete;
           expect(parent.wizardUI.inputs[1].checkValidity()).to.be.true;
         })
