+++ test/integration/editors/substation/guess-wizarding-editing.test.ts
@@ -1,15 +1,15 @@
 import { expect, fixture, html } from '@open-wc/testing';
 
-import '../../mock-wizard-editor.js';
+import '../../../mock-wizard-editor.js';
 
 import '@material/mwc-list/mwc-check-list-item';
 import '@material/mwc-list/mwc-list';
 import '../../mock-wizard.js';
-import { getDocument } from '../../data.js';
-import { MockWizard } from '../../mock-wizard.js';
-import { guessVoltageLevel } from '../../../src/editors/substation/guess-wizard.js';
-import { EditingElement } from '../../../src/Editing.js';
-import { WizardingElement } from '../../../src/Wizarding.js';
+import { getDocument } from '../../../data.js';
+import { MockWizard } from '../../../mock-wizard.js';
+import { guessVoltageLevel } from '../../../../src/editors/substation/guess-wizard.js';
+import { EditingElement } from '../../../../src/Editing.js';
+import { WizardingElement } from '../../../../src/Wizarding.js';
 
 describe('guess-wizard-integration', () => {
   let element: MockWizard;
