+++ test/integration/editors/substation/substation-editor-wizarding-editing.test.ts
@@ -1,13 +1,13 @@
 import { fixture, html, expect } from '@open-wc/testing';
 
-import '../../mock-wizard-editor.js';
-import { EditingElement } from '../../../src/Editing.js';
-import { WizardingElement } from '../../../src/Wizarding.js';
+import '../../../mock-wizard-editor.js';
+import { EditingElement } from '../../../../src/Editing.js';
+import { WizardingElement } from '../../../../src/Wizarding.js';
 
-import { getDocument } from '../../data.js';
+import { getDocument } from '../../../data.js';
 
 import { ListItemBase } from '@material/mwc-list/mwc-list-item-base';
-import { SubstationEditor } from '../../../src/editors/substation/substation-editor.js';
+import { SubstationEditor } from '../../../../src/editors/substation/substation-editor.js';
 
 describe('substation-editor wizarding editing integration', () => {
   describe('edit wizard', () => {
