+++ test/integration/editors/substation/voltage-level-editor-wizarding.test.ts
@@ -1,11 +1,11 @@
 import { fixture, html, expect } from '@open-wc/testing';
 import fc from 'fast-check';
 
-import '../../mock-wizard.js';
-import { WizardingElement } from '../../../src/Wizarding.js';
+import '../../../mock-wizard.js';
+import { WizardingElement } from '../../../../src/Wizarding.js';
 
-import { getDocument } from '../../data.js';
-import { regexString, regExp, inverseRegExp } from '../../foundation.js';
+import { getDocument } from '../../../data.js';
+import { regexString, regExp, inverseRegExp } from '../../../foundation.js';
 
 describe('voltage-level-editor wizarding integration', () => {
   const doc = getDocument();
